feat: CoreData 생성 및 textView에 키보드 기능추가
