+++ Diary+CoreDataClass.swift
@@ -1,0 +1,15 @@
+//
+//  Diary+CoreDataClass.swift
+//  Diary
+//
+//  Created by 예찬 on 2023/09/05.
+//
+//
+
+import Foundation
+import CoreData
+
+@objc(Diary)
+public class Diary: NSManagedObject {
+
+}
