+++ Diary/Error/DecodingError.swift
@@ -5,10 +5,12 @@
 //  Created by idinaloq, yetti on 2023/08/30.
 //
 
-enum DecodingError: Error {
+import Foundation
+
+enum DecodingError: LocalizedError {
     case decodingFailure
     
-    var description: String {
+    var errorDescription: String? {
         switch self {
         case .decodingFailure:
             return "디코딩 에러입니다."
