+++ Diary/Manager/CoreDataManager.swift
@@ -1,0 +1,9 @@
+//
+//  CoreDataManager.swift
+//  Diary
+//
+//  Created by 예찬 on 2023/09/04.
+//
+
+import UIKit
+import CoreData
