+++ Diary/Resource/Info.plist
